Test: InputViewTest에서 이루어진 검증 부분 테스트 코드를 ValidatorTest에서 작성
